requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
tzdata==2025.2
mysqlclient==2.2.7
python-dotenv
djangorestframework==3.15.2
django-cors-headers